+++ fantoch_ps/src/protocol/epaxos.rs
@@ -40,7 +40,7 @@ impl<KC: KeyClocks> Protocol for EPaxos<KC> {
     fn new(
         process_id: ProcessId,
         config: Config,
-    ) -> (Self, Vec<(Self::PeriodicEvent, u64)>) {
+    ) -> (Self, Vec<(Self::PeriodicEvent, usize)>) {
         // compute fast and write quorum sizes
         let (fast_quorum_size, write_quorum_size) =
             config.epaxos_quorum_sizes();
